Consider the extensions for the main meny in the standalone mode
