+++ web_app.py
@@ -1,8 +1,6 @@
-from importlib.metadata import entry_points
-
 from flask import g
 
-from app_context.appmenu import ApplicationMenu
+from app_context.appmenu import compose_menu
 from flask_app import init_app
 
 app = init_app()
@@ -10,26 +8,10 @@ app = init_app()
 
 @app.before_request
 def init_app_menu():
-    menu = ApplicationMenu()
-
-    menu.add_section("vod_chat", "VOD Chat", target_url="/vod-chat/")
-
-    _load_menu_extensions(menu)
+    menu = compose_menu()
 
     # TODO Cache the value
     g.main_menu = menu
-
-
-def _load_menu_extensions(menu: ApplicationMenu) -> None:
-    discovered_extensions = entry_points(group="chat_analyzer.v1.blueprints", name="inject_menu")
-
-    for extension in sorted(discovered_extensions):
-        try:
-            inject_menu = extension.load()
-            inject_menu(menu)
-        except Exception:
-            print(f"Failed to update main manu by '{extension.module}' extension", flush=True)
-            raise
 
 
 if __name__ == '__main__':
